wmediumd: port to usfstl
Port to the usfstl for time/scheduling, which should also fix an
issue with station moving not happening unless there are frames
in each move interval (which is likely since it's long, so the
issue is likely only theoretical).

Also add time integration (sched_ctrl) using the user-mode linux
protocol.

Tested with hwsim tests from wpa_supplicant (in user-mode linux
with time-travel) as well as with actually a time-control app
that I'm going to release soon.